pyproject.toml
[project]
name = "psio-disc-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pillow>=11.3.0",
    "requests>=2.32.4",
]
